Support #24914 - Add support for type link
- externalResourceURI has been removed from the Derivative objects.
- isExternal is now not stored in the database.
- Changeset has been changed to not include isExternal and everything from the derivative.
- DerivativeValidator has been reverted.
