+++ src/main/java/ca/gc/aafc/objectstore/api/dto/ObjectStoreMetadataDto.java
@@ -56,7 +56,6 @@ public class ObjectStoreMetadataDto extends AttributeMetaInfoProvider {
   private String bucket;
   private UUID fileIdentifier;
   private String fileExtension;
-  private Boolean isExternal;
   private String resourceExternalURI;
 
   private String dcFormat;
